+++ modules/base_infra/log_archive/s3_logs_archive.tf
@@ -255,7 +255,7 @@ resource "aws_s3_bucket_public_access_block" "s3_logs_archive_bucket_public_acce
 resource "null_resource" "cleanup_s3_logs_archive_bucket" {
   triggers = {
     bucket_name = aws_s3_bucket.s3_logs_archive_bucket.id
-    region      = data.aws_region.current.name
+    region      = data.aws_region.current.region
     account     = data.aws_caller_identity.current.account_id
   }