+++ modules/cirrus/workflow/main.tf
@@ -4,7 +4,7 @@ data "aws_region" "current" {}
 locals {
   # Save as locals to avoid obnoxiously long lines
   current_account = data.aws_caller_identity.current.account_id
-  current_region  = data.aws_region.current.name
+  current_region  = data.aws_region.current.region
 
   # Create the workflow's state machine JSON.
   # Use the Cirrus task output map, user-defined template variable map, and