+++ modules/mosaic-titiler/titler-mosaicjson-private-api.tf
@@ -1,5 +1,15 @@
 locals {
   name_prefix = "fd-${var.project_name}-${var.environment}"
+  # if var.custom_vpce_id was provided, the user is indicating they want to use a custom vpc endpoint which they have
+  # defined. if not, and if is_private_endpoint is true, we use the vpce we create in this module
+  # vpce_id = try(coalesce(var.custom_vpce_id, aws_vpc_endpoint.titiler_api_gateway_private[0].id), null)
+  vpce_id = var.custom_vpce_id != null ? var.custom_vpce_id : (var.is_private_endpoint ? aws_vpc_endpoint.titiler_api_gateway_private[0].id : null)
+  # additionally, only create a vpc endpoint in this module if the api gateway is private *and* the user has not
+  # indicated they are using their own vpc endpoint
+  create_vpce = var.is_private_endpoint == true && var.custom_vpce_id == null
+
+  // ensures we use the same var everywhere stage_name of the gateway is needed, and helps avoid ciricular deps
+  stage_name = var.environment
 }
 
 resource "aws_security_group" "titiler_api_gateway_private_vpce" {
@@ -24,15 +34,15 @@ resource "aws_vpc_security_group_ingress_rule" "titiler_api_gateway_private_vcpe
 }
 
 resource "aws_vpc_endpoint" "titiler_api_gateway_private" {
-  count = var.is_private_endpoint ? 1 : 0
+  count = local.create_vpce ? 1 : 0
 
-  service_name        = "com.amazonaws.${data.aws_region.current.name}.execute-api"
+  service_name        = "com.amazonaws.${data.aws_region.current.region}.execute-api"
   vpc_id              = var.vpc_id
   vpc_endpoint_type   = "Interface"
   ip_address_type     = "ipv4"
   subnet_ids          = data.aws_subnet.selected[*].id
   auto_accept         = true
-  private_dns_enabled = false
+  private_dns_enabled = var.vpce_private_dns_enabled
   security_group_ids = concat(
     aws_security_group.titiler_api_gateway_private_vpce[*].id,
     coalesce(var.private_api_additional_security_group_ids, [])
@@ -47,9 +57,13 @@ resource "aws_api_gateway_rest_api" "titiler_api_gateway" {
   count = var.is_private_endpoint ? 1 : 0
   name  = "${local.name_prefix}-titiler"
 
+  # titiler serves binary data (images), thus the need to enable binary media types. without this, api gateway may
+  # return the binary data as plain text, breaking titiler's functionality
+  binary_media_types = ["image/*", "application/octet-stream"]
+
   endpoint_configuration {
     types            = ["PRIVATE"]
-    vpc_endpoint_ids = var.is_private_endpoint ? aws_vpc_endpoint.titiler_api_gateway_private[*].id : null
+    vpc_endpoint_ids = var.is_private_endpoint ? [local.vpce_id] : null
   }
 }
 
@@ -60,7 +74,7 @@ data "aws_iam_policy_document" "titiler_api_gateway_private" {
     sid       = "DenyApiInvokeForNonVpceTraffic"
     effect    = "Deny"
     actions   = ["execute-api:Invoke"]
-    resources = ["arn:aws:execute-api:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:${aws_api_gateway_rest_api.titiler_api_gateway[0].id}/*"]
+    resources = ["arn:aws:execute-api:${data.aws_region.current.region}:${data.aws_caller_identity.current.account_id}:${aws_api_gateway_rest_api.titiler_api_gateway[0].id}/*"]
 
     principals {
       type        = "AWS"
@@ -70,7 +84,7 @@ data "aws_iam_policy_document" "titiler_api_gateway_private" {
     condition {
       variable = "aws:SourceVpce"
       test     = "StringNotEquals"
-      values   = [aws_vpc_endpoint.titiler_api_gateway_private[0].id]
+      values   = [local.vpce_id]
     }
   }
 
@@ -78,7 +92,7 @@ data "aws_iam_policy_document" "titiler_api_gateway_private" {
     sid       = "AllowApiInvoke"
     effect    = "Allow"
     actions   = ["execute-api:Invoke"]
-    resources = ["arn:aws:execute-api:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:${aws_api_gateway_rest_api.titiler_api_gateway[0].id}/*"]
+    resources = ["arn:aws:execute-api:${data.aws_region.current.region}:${data.aws_caller_identity.current.account_id}:${aws_api_gateway_rest_api.titiler_api_gateway[0].id}/*"]
 
     principals {
       type        = "AWS"
@@ -108,7 +122,7 @@ resource "aws_api_gateway_integration" "titiler_api_gateway_root_method_integrat
   resource_id             = aws_api_gateway_rest_api.titiler_api_gateway[0].root_resource_id
   http_method             = aws_api_gateway_method.titiler_api_gateway_root_method[0].http_method
   type                    = "AWS_PROXY"
-  uri                     = "arn:aws:apigateway:${data.aws_region.current.name}:lambda:path/2015-03-31/functions/${aws_lambda_function.titiler-mosaic-lambda.arn}/invocations"
+  uri                     = "arn:aws:apigateway:${data.aws_region.current.region}:lambda:path/2015-03-31/functions/${aws_lambda_function.titiler-mosaic-lambda.arn}/invocations"
   integration_http_method = "POST"
 }
 
@@ -217,7 +231,7 @@ resource "aws_api_gateway_integration" "titiler_api_gateway_proxy_resource_metho
   resource_id             = aws_api_gateway_resource.titiler_api_gateway_proxy_resource[0].id
   http_method             = aws_api_gateway_method.titiler_api_gateway_proxy_resource_method[0].http_method
   type                    = "AWS_PROXY"
-  uri                     = "arn:aws:apigateway:${data.aws_region.current.name}:lambda:path/2015-03-31/functions/${aws_lambda_function.titiler-mosaic-lambda.arn}/invocations"
+  uri                     = "arn:aws:apigateway:${data.aws_region.current.region}:lambda:path/2015-03-31/functions/${aws_lambda_function.titiler-mosaic-lambda.arn}/invocations"
   integration_http_method = "POST"
 }
 
@@ -228,44 +242,32 @@ resource "aws_api_gateway_deployment" "titiler_api_gateway" {
     aws_api_gateway_integration.titiler_api_gateway_proxy_resource_method_integration,
   ]
 
-  rest_api_id       = aws_api_gateway_rest_api.titiler_api_gateway[0].id
-  stage_name        = var.environment
-  stage_description = var.titiler_api_stage_description
+  rest_api_id = aws_api_gateway_rest_api.titiler_api_gateway[0].id
 
   lifecycle {
     create_before_destroy = true
   }
 }
 
+resource "aws_api_gateway_stage" "titiler_api_gateway_stage" {
+  count = var.is_private_endpoint ? 1 : 0
+
+  deployment_id = aws_api_gateway_deployment.titiler_api_gateway[0].id
+  rest_api_id   = aws_api_gateway_rest_api.titiler_api_gateway[0].id
+  stage_name    = local.stage_name
+  description   = var.titiler_api_stage_description
+
+  access_log_settings {
+    destination_arn = aws_cloudwatch_log_group.titiler_api_gateway_logs_group[0].arn
+    # terraform's jsonencode() sorts keys lexicographically, which would modify the log format. so, we build a string
+    # https://github.com/hashicorp/terraform/issues/27880
+    format = "{requestId:$context.requestId,ip:$context.identity.sourceIp,caller:$context.identity.caller,user:$context.identity.user,requestTime:$context.requestTime,httpMethod:$context.httpMethod,resourcePath:$context.resourcePath,status:$context.status,protocol:$context.protocol,responseLength:$context.responseLength}"
+  }
+}
+
 resource "aws_cloudwatch_log_group" "titiler_api_gateway_logs_group" {
   count = var.is_private_endpoint ? 1 : 0
-  name  = "/aws/apigateway/${local.name_prefix}-titiler-${aws_api_gateway_deployment.titiler_api_gateway[0].rest_api_id}/${aws_api_gateway_deployment.titiler_api_gateway[0].stage_name}"
-}
-
-locals {
-  access_log_format = "{\"requestId\":\"\\$context.requestId\",\"ip\":\"\\$context.identity.sourceIp\",\"caller\":\"\\$context.identity.caller\",\"user\":\"\\$context.identity.user\",\"requestTime\":\"\\$context.requestTime\",\"httpMethod\":\"\\$context.httpMethod\",\"resourcePath\":\"\\$context.resourcePath\",\"status\":\"\\$context.status\",\"protocol\":\"\\$context.protocol\",\"responseLength\":\"\\$context.responseLength\"}"
-}
-
-resource "null_resource" "enable_access_logs" {
-  count = var.is_private_endpoint ? 1 : 0
-  triggers = {
-    stage_name              = aws_api_gateway_deployment.titiler_api_gateway[0].stage_name
-    rest_api_id             = aws_api_gateway_deployment.titiler_api_gateway[0].rest_api_id
-    apigw_access_logs_group = aws_cloudwatch_log_group.titiler_api_gateway_logs_group[0].arn
-    access_log_format       = local.access_log_format
-  }
-
-  provisioner "local-exec" {
-    interpreter = ["bash", "-ec"]
-    command     = <<EOF
-export AWS_DEFAULT_REGION=${data.aws_region.current.name}
-export AWS_REGION=${data.aws_region.current.name}
-
-echo "Update Access Logging on FilmDrop TiTiler API."
-aws apigateway update-stage --rest-api-id ${aws_api_gateway_deployment.titiler_api_gateway[0].rest_api_id} --stage-name ${aws_api_gateway_deployment.titiler_api_gateway[0].stage_name} --patch-operations "[{\"op\": \"replace\",\"path\": \"/accessLogSettings/destinationArn\",\"value\": \"${aws_cloudwatch_log_group.titiler_api_gateway_logs_group[0].arn}\"},{\"op\": \"replace\",\"path\": \"/accessLogSettings/format\",\"value\": \"${local.access_log_format}\"}]"
-
-EOF
-  }
+  name  = "/aws/apigateway/${local.name_prefix}-titiler-${aws_api_gateway_deployment.titiler_api_gateway[0].rest_api_id}/${local.stage_name}"
 }
 
 resource "aws_lambda_permission" "titiler_api_gateway_lambda_permission_root_resource" {
@@ -275,7 +277,7 @@ resource "aws_lambda_permission" "titiler_api_gateway_lambda_permission_root_res
   function_name = aws_lambda_function.titiler-mosaic-lambda.arn
   principal     = "apigateway.amazonaws.com"
 
-  source_arn = "arn:aws:execute-api:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:${aws_api_gateway_rest_api.titiler_api_gateway[0].id}/*/*"
+  source_arn = "arn:aws:execute-api:${data.aws_region.current.region}:${data.aws_caller_identity.current.account_id}:${aws_api_gateway_rest_api.titiler_api_gateway[0].id}/*/*"
 }
 
 resource "aws_api_gateway_domain_name" "titiler_api_gateway_domain_name" {
@@ -295,16 +297,16 @@ resource "aws_api_gateway_domain_name" "titiler_api_gateway_domain_name" {
       "Effect": "Allow",
       "Principal": "*",
       "Action": "execute-api:Invoke",
-      "Resource": "arn:aws:execute-api:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:/domainnames/*"
+      "Resource": "arn:aws:execute-api:${data.aws_region.current.region}:${data.aws_caller_identity.current.account_id}:/domainnames/*"
     },
     {
       "Effect": "Deny",
       "Principal": "*",
       "Action": "execute-api:Invoke",
-      "Resource": "arn:aws:execute-api:${data.aws_region.current.name}:${data.aws_caller_identity.current.account_id}:/domainnames/*",
+      "Resource": "arn:aws:execute-api:${data.aws_region.current.region}:${data.aws_caller_identity.current.account_id}:/domainnames/*",
       "Condition": {
         "StringNotEquals": {
-          "aws:SourceVpce": "${aws_vpc_endpoint.titiler_api_gateway_private[0].id}"
+          "aws:SourceVpce": "${local.vpce_id}"
         }
       }
     }
@@ -315,7 +317,7 @@ EOF
 
 resource "aws_api_gateway_domain_name_access_association" "titiler_api_gateway_domain_name_access_association" {
   count                          = var.is_private_endpoint == true && var.domain_alias != "" && var.private_certificate_arn != "" ? 1 : 0
-  access_association_source      = aws_vpc_endpoint.titiler_api_gateway_private[0].id
+  access_association_source      = local.vpce_id
   access_association_source_type = "VPCE"
   domain_name_arn                = aws_api_gateway_domain_name.titiler_api_gateway_domain_name[0].arn
 }
@@ -325,5 +327,5 @@ resource "aws_api_gateway_base_path_mapping" "titiler_api_gateway_domain_mapping
   domain_name    = aws_api_gateway_domain_name.titiler_api_gateway_domain_name[0].domain_name
   domain_name_id = aws_api_gateway_domain_name.titiler_api_gateway_domain_name[0].domain_name_id
   api_id         = aws_api_gateway_rest_api.titiler_api_gateway[0].id
-  stage_name     = aws_api_gateway_deployment.titiler_api_gateway[0].stage_name
+  stage_name     = local.stage_name
 }