+++ CHANGELOG.md
@@ -11,12 +11,29 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- A new stac_server_inputs.custom_vpce_id var added. If provided, the user is indicating that they have an existing vpc endpoint that the stac server api gateway (and supporting resources) should allow to ingress
+
+- A new titiler_inputs.custom_vpce_id var added. If provided, the user is indicating that they have an existing vpc endpoint that the titiler api gateway (and supporting resources) should allow to ingress
+
 ### Changed
+
+- Minor changes ([#218](https://github.com/Element84/filmdrop-aws-tf-modules/pull/218)) to how TiTiler behaves when is_private_endpoint = true, supporting communications with its API Gateway and VPC Endpoint
+
+- TiTiler module will note three IAM related resources being created: aws_iam_role_policy.titiler-mosaic-lambda-inline-policy, aws_iam_role_policy_attachment.lambda_basic_execution, aws_iam_role_policy_attachment.lambda_vpc_access. This is a result of AWS provider deprecations to managed_policy and inline_policy properties. The TiTiler role in question will updated in place, resulting in no material changes to the role.
 
 ### Fixed
 
+- Corrected permissions in Workflow Metrics CloudWatch write policy to include
+  necessary wildcard, and in the read policy to include listing
+  metrics. ([#237](https://github.com/Element84/filmdrop-aws-tf-modules/pull/237))
+
 ### Removed
 
+- If you have stac server, cirrus, or titiler (as *private*) deployed`, Terraform will note that a null_resource.enable_access_logs is being destroyed for each. This is expected. Previously, api gateway access logs were enabled via this null resource; they are now enabled via an aws_api_gateway_stage resource.
+
+### ⚠️ Breaking
+
+- See MIGRATION.md for important updates to the AWS provider version to v6, and the terraform-aws-stac-server module to v2